requirements.txt
annotated-types==0.7.0
anyio==4.9.0
aiohttp>=3.8.0
certifi==2025.7.14
click>=8.1.0
colorama==0.4.6
deprecation==2.1.0
distro==1.9.0
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.116.1
gotrue==2.12.3
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
jiter==0.10.0
mangum==0.19.0
openai==1.97.0
packaging==25.0
postgrest==1.1.1
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart>=0.0.5
realtime==2.6.0
six==1.17.0
sniffio==1.3.1
starlette==0.47.1
storage3==0.12.0
StrEnum==0.4.15
supabase==2.17.0
supafunc==0.10.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
uvicorn==0.35.0
websockets==15.0.1
PyPDF2==3.0.1
python-docx==1.2.0
autopep8==2.3.2
stripe==10.5.0
resend>=2.23.0
